Separated wcAPI into a new file for reuse
